+++ src/main/java/com/github/johnnyjayjay/discord/commandapi/CommandEvent.java
@@ -14,15 +14,18 @@ import java.util.Optional;
 /**
  * Represents a command event. This is not much different from a GuildMessageReceivedEvent, though it gives access to the called command.
  * @author Johnny_JayJay
- * @version 3.1_1
+ * @version 3.2
+ * @see GuildMessageReceivedEvent
  */
 public class CommandEvent extends GuildMessageReceivedEvent {
 
     private final Command command;
+    private final CommandSettings settings;
 
-    public CommandEvent(JDA api, long responseNumber, Message message, Command command) {
+    public CommandEvent(JDA api, long responseNumber, Message message, Command command, CommandSettings settings) {
         super(api, responseNumber, message);
         this.command = command;
+        this.settings = settings;
     }
 
     /**
@@ -30,7 +33,7 @@ public class CommandEvent extends GuildMessageReceivedEvent {
      * @param msg The message to respond with as a String.
      */
     public void respond(String msg) {
-        if (checkBotPermissions(Permission.MESSAGE_WRITE))
+        if (checkBotPermissions(Permission.MESSAGE_WRITE, Permission.MESSAGE_EMBED_LINKS))
             this.getChannel().sendMessage(msg).queue();
     }
 
@@ -39,7 +42,7 @@ public class CommandEvent extends GuildMessageReceivedEvent {
      * @param msg The message to respond with as a MessageEmbed.
      */
     public void respond(MessageEmbed msg) {
-        if (checkBotPermissions(Permission.MESSAGE_WRITE))
+        if (checkBotPermissions(Permission.MESSAGE_WRITE, Permission.MESSAGE_EMBED_LINKS))
             this.getChannel().sendMessage(msg).queue();
     }
 
@@ -48,7 +51,7 @@ public class CommandEvent extends GuildMessageReceivedEvent {
      * @param msg The message to respond with as a Message object.
      */
     public void respond(Message msg) {
-        if (checkBotPermissions(Permission.MESSAGE_WRITE))
+        if (checkBotPermissions(Permission.MESSAGE_WRITE, Permission.MESSAGE_EMBED_LINKS))
             this.getChannel().sendMessage(msg).queue();
     }
 
@@ -58,6 +61,15 @@ public class CommandEvent extends GuildMessageReceivedEvent {
      */
     public Command getCommand() {
         return command;
+    }
+
+    /**
+     * Returns the CommandSettings instance this command was called for.
+     * @return an instance of CommandSettings.
+     * @see CommandSettings
+     */
+    public CommandSettings getCommandSettings() {
+        return settings;
     }
 
     /**
@@ -111,7 +123,7 @@ public class CommandEvent extends GuildMessageReceivedEvent {
 
     /**
      * Returns whether the Member who executed the command has the given permissions in the event channel.
-     * @param permissions One or more Permissions to check for.
+     * @param permissions One or more Checks to check for.
      * @return true: Member has the permissions in this channel, false: they have not
      */
     public boolean checkMemberPermissions(Permission... permissions) {
@@ -120,7 +132,7 @@ public class CommandEvent extends GuildMessageReceivedEvent {
 
     /**
      * Returns whether the self member has the given permissions in the event channel.
-     * @param permissions One or more Permissions to check for.
+     * @param permissions One or more Checks to check for.
      * @return true: self member has the permissions in this channel, false: it has not
      */
     public boolean checkBotPermissions(Permission... permissions) {