+++ changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-### 3.1_1
+### 3.2
 - Updated to JDA version 3.7.1_387
 - Adjusted cooldown system: it is now configurable whether the cooldown will reset for each command execution attempt
 - Deprecated help label methods: They will be removed in the near future as the help label system will be merged with the normal command system
@@ -13,6 +13,13 @@
 - Added method `CommandSettings#getLabels(ICommand)` to get a Set of labels for a specific command
 - Added more fields to `CommandEvent.Command`
 - Added method `isBlacklisted(long)` to `CommandSettings`
+- Added classes AbstractHelpCommand and DefaultHelpCommand
+- Added method `getCommandSettings` in `CommandEvent`
+- Annotated `ICommand` as a `FunctionalInterface`
+- Removed supported for help label methods in `CommandSettings`
+- Cleaned `CommandListener`
+- Deprecated method `String info(Member)` in `ICommand`;
+    New method is `Message info(Member, String, Set<String>)` because it is more flexible.
 
 ### 3.1
 - Changed setCooldown-method's return to `this`